IR sensors are analog not digital
Changes the command used to read the IR sensors to analogRead as necessary for the IRsensor circuitry.  The analog approach is also simpler for a student to understand because it is simply the conversion of a variable voltage to a 10-bit integer (a number from 0 to 1023) that corresponds to the voltage.  The digital approach requires circuitry with a capacitor and requires writing code to capture how long it takes the capacitor to switch from high to low voltage.